requirements.txt
cryptography
pytest
pynacl>=1.5.0
